requirements.txt
aenum==3.1.15
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
asgiref==3.2.10
attrs==23.2.0
blis==0.7.11
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6
confection==0.1.5
configobj==5.0.8
configparser==7.0.0
cymem==2.0.8
Cython==3.0.10
distlib==0.3.8
distro==1.9.0
dj-database-url==0.5.0
Django==3.1.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
es-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.7.0/es_core_news_sm-3.7.0-py3-none-any.whl#sha256=61e6e5530941f5880166855f09f60d7e6ba79ec1e8e45f96244bdb1eb169eb1d
etelemetry==0.3.1
filelock==3.13.1
fitz==0.0.1.dev2
frozenlist==1.4.1
fsspec==2024.6.0
google-ai-generativelanguage==0.6.5
google-api-core==2.19.0
google-api-python-client==2.133.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.0
googleapis-common-protos==1.63.1
gradientai==1.13.0
grpcio==1.64.1
grpcio-status==1.62.2
gunicorn==20.0.4
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.6
intel-openmp==2021.4.0
isodate==0.6.1
Jinja2==3.1.4
joblib==1.3.2
langcodes==3.4.0
langdetect==1.0.9
language_data==1.2.0
looseversion==1.3.0
lxml==5.2.2
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
mysql-connector==2.2.9
neattext==0.1.3
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
nltk==3.8.1
numpy==1.26.4
openai==0.28.0
packaging==24.1
pandas==2.2.2
pathlib==1.0.1
pillow==10.2.0
preshed==3.0.9
proto-plus==1.23.0
protobuf==4.25.3
prov==2.0.1
psycopg2==2.8.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.4
pydantic_core==2.18.4
pydot==2.0.0
Pygments==2.18.0
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
pyparsing==3.1.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.1
rdflib==6.3.2
regex==2024.5.15
requests==2.31.0
rich==13.7.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.4.1.post1
scipy==1.12.0
setuptools==69.1.1
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.4
srsly==2.4.8
sympy==1.12.1
tbb==2021.12.0
textblob==0.18.0.post0
thinc==8.2.5
threadpoolctl==3.3.0
tokenizers==0.19.1
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tqdm==4.66.4
traits==6.3.2
transformers==4.41.2
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.0.31
wasabi==1.1.3
weasel==0.4.1
whitenoise==5.2.0
wrapt==1.16.0
yarl==1.9.4
